test: coverage cli

+++ cli.cjs
@@ -1,189 +1,188 @@
 #!/usr/bin/env bun
-const { spawnSync } = require("child_process");
+const cp = require("child_process");
 const path = require("node:path");
 const fs = require("node:fs");
-const packageJSON = await import(path.join(process.cwd(), "package.json")).then(
-  (m) => m.default,
-);
+const outPath = path.join(import.meta.dir, 'out');
 
-const BRISA_BUILD_FOLDER =
-  process.env.BRISA_BUILD_FOLDER || path.join(process.cwd(), "build");
+export async function main() {
+  const packageJSON = require(path.join(process.cwd(), "package.json"));
+  const BRISA_BUILD_FOLDER =
+    process.env.BRISA_BUILD_FOLDER || path.join(process.cwd(), "build");
 
-const prodOptions = {
-  stdio: "inherit",
-  env: { ...process.env, NODE_ENV: "production", BRISA_BUILD_FOLDER },
-};
-const devOptions = {
-  stdio: "inherit",
-  env: { ...process.env, NODE_ENV: "development", BRISA_BUILD_FOLDER },
-};
+  const prodOptions = {
+    stdio: "inherit",
+    env: { ...process.env, NODE_ENV: "production", BRISA_BUILD_FOLDER },
+  };
+  const devOptions = {
+    stdio: "inherit",
+    env: { ...process.env, NODE_ENV: "development", BRISA_BUILD_FOLDER },
+  };
 
-let BUN_EXEC;
-let BUNX_EXEC;
-let IS_DESKTOP_APP = false; // default value depends on brisa.config.ts
+  let BUN_EXEC;
+  let BUNX_EXEC;
+  let IS_DESKTOP_APP = false; // default value depends on brisa.config.ts
 
-// Check if is desktop app
-try {
-  const config = await import(path.join(process.cwd(), "brisa.config.ts")).then(
-    (m) => m.default,
-  );
+  // Check if is desktop app
+  try {
+    const config = require(path.join(process.cwd(), "brisa.config.ts")).default;
 
-  IS_DESKTOP_APP =
-    typeof config.output === "string" && config.output === "desktop";
-} catch (error) {}
+    IS_DESKTOP_APP =
+      typeof config.output === "string" && config.output === "desktop";
+  } catch (error) {}
 
-try {
-  // Check if 'bun' is available in the system
-  const bunCheck = spawnSync("bun", ["--version"], { stdio: "ignore" });
-  if (bunCheck.status === 0) {
-    BUN_EXEC = "bun";
-    BUNX_EXEC = "bunx";
-  } else {
-    BUN_EXEC = `${process.env.HOME}/.bun/bin/bun`;
-    BUNX_EXEC = `${process.env.HOME}/.bun/bin/bunx`;
-  }
+  try {
+    // Check if 'bun' is available in the system
+    const bunCheck = cp.spawnSync("bun", ["--version"], { stdio: "ignore" });
+    if (bunCheck.status === 0) {
+      BUN_EXEC = "bun";
+      BUNX_EXEC = "bunx";
+    } else {
+      BUN_EXEC = `${process.env.HOME}/.bun/bin/bun`;
+      BUNX_EXEC = `${process.env.HOME}/.bun/bin/bunx`;
+    }
 
-  // Command: brisa dev
-  if (process.argv[2] === "dev") {
-    let PORT = 3000; // default port
-    let DEBUG_MODE = false; // default debug mode
+    // Command: brisa dev
+    if (process.argv[2] === "dev") {
+      let PORT = 3000; // default port
+      let DEBUG_MODE = false; // default debug mode
 
-    for (let i = 3; i < process.argv.length; i++) {
-      switch (process.argv[i]) {
-        case "--skip-desktop":
-        case "-s":
-          IS_DESKTOP_APP = false;
-          break;
-        case "-p":
-        case "--port":
-          PORT = process.argv[i + 1];
-          i++;
-          break;
-        case "-d":
-        case "--debug":
-          DEBUG_MODE = true;
-          break;
-        case "--help":
-          console.log("Usage: brisa dev [options]");
-          console.log("Options:");
-          console.log(" -p, --port         Specify port");
-          console.log(" -d, --debug        Enable debug mode");
-          console.log(
-            " -s, --skip-desktop Skip open desktop app when 'output': 'desktop' in brisa.config.ts",
-          );
-          console.log(" --help             Show help");
-          process.exit(0);
+      for (let i = 3; i < process.argv.length; i++) {
+        switch (process.argv[i]) {
+          case "--skip-desktop":
+          case "-s":
+            IS_DESKTOP_APP = false;
+            break;
+          case "-p":
+          case "--port":
+            PORT = process.argv[i + 1];
+            i++;
+            break;
+          case "-d":
+          case "--debug":
+            DEBUG_MODE = true;
+            break;
+          case "--help":
+            console.log("Usage: brisa dev [options]");
+            console.log("Options:");
+            console.log(" -p, --port         Specify port");
+            console.log(" -d, --debug        Enable debug mode");
+            console.log(
+              " -s, --skip-desktop Skip open desktop app when 'output': 'desktop' in brisa.config.ts",
+            );
+            console.log(" --help             Show help");
+            return process.exit(0);
+        }
+      }
+
+      const buildCommand = [path.join(outPath, 'cli', 'build.js'), "DEV"];
+      const serveCommand = [
+        path.join(outPath, "cli", "serve", "index.js"),
+        `${PORT}`,
+        "DEV",
+      ];
+
+      // DEV mode for desktop app
+      if (IS_DESKTOP_APP) {
+        await initTauri(devOptions, PORT);
+        cp.spawnSync(BUNX_EXEC, ["tauri", "dev", "--port", PORT.toString()], devOptions);
+      } else if (DEBUG_MODE) {
+        cp.spawnSync(BUN_EXEC, buildCommand, devOptions);
+        cp.spawnSync(BUN_EXEC, ["--inspect", ...serveCommand], devOptions);
+      } else {
+        cp.spawnSync(BUN_EXEC, buildCommand, devOptions);
+        cp.spawnSync(BUN_EXEC, serveCommand, devOptions);
       }
     }
 
-    const buildCommand = ["node_modules/brisa/out/cli/build.js", "DEV"];
-    const serveCommand = [
-      "node_modules/brisa/out/cli/serve/index.js",
-      `${PORT}`,
-      "DEV",
-    ];
+    // Command: brisa build
+    else if (process.argv[2] === "build") {
+      for (let i = 3; i < process.argv.length; i++) {
+        switch (process.argv[i]) {
+          case "--skip-desktop":
+          case "-s":
+            IS_DESKTOP_APP = false;
+            break;
+          case "--help":
+            console.log("Usage: brisa build [options]");
+            console.log("Options:");
+            console.log(
+              " -s, --skip-desktop Skip open desktop app when 'output': 'desktop' in brisa.config.ts",
+            );
+            console.log(" --help             Show help");
+            return process.exit(0);
+        }
+      }
 
-    // DEV mode for desktop app
-    if (IS_DESKTOP_APP) {
-      await initTauri();
-      spawnSync(BUNX_EXEC, "tauri dev --port 3000".split(" "), devOptions);
-    } else if (DEBUG_MODE) {
-      spawnSync(BUN_EXEC, buildCommand, devOptions);
-      spawnSync(BUN_EXEC, ["--inspect", ...serveCommand], devOptions);
-    } else {
-      spawnSync(BUN_EXEC, buildCommand, devOptions);
-      spawnSync(BUN_EXEC, serveCommand, devOptions);
-    }
-  }
-
-  // Command: brisa build
-  else if (process.argv[2] === "build") {
-    for (let i = 3; i < process.argv.length; i++) {
-      switch (process.argv[i]) {
-        case "--skip-desktop":
-        case "-s":
-          IS_DESKTOP_APP = false;
-          break;
-        case "--help":
-          console.log("Usage: brisa build [options]");
-          console.log("Options:");
-          console.log(
-            " -s, --skip-desktop Skip open desktop app when 'output': 'desktop' in brisa.config.ts",
-          );
-          console.log(" --help             Show help");
-          process.exit(0);
+      if (IS_DESKTOP_APP) {
+        await initTauri(prodOptions);
+        cp.spawnSync(BUNX_EXEC, ["tauri", "build"], prodOptions);
+      } else {
+        cp.spawnSync(
+          BUN_EXEC,
+          [path.join(outPath, 'cli', 'build.js'), "PROD"],
+          prodOptions,
+        );
       }
     }
 
-    if (IS_DESKTOP_APP) {
-      await initTauri();
-      spawnSync(BUNX_EXEC, ["tauri", "build"], devOptions);
-    } else {
-      spawnSync(
+    // Command: brisa start
+    else if (process.argv[2] === "start") {
+      let PORT = 3000; // default port
+
+      for (let i = 3; i < process.argv.length; i++) {
+        switch (process.argv[i]) {
+          case "-p":
+          case "--port":
+            PORT = process.argv[i + 1];
+            i++;
+            break;
+          case "--help":
+            console.log("Usage: brisa start [options]");
+            console.log("Options:");
+            console.log(" -p, --port    Specify port");
+            console.log(" --help        Show help");
+            return process.exit(0);
+        }
+      }
+
+      cp.spawnSync(
         BUN_EXEC,
-        ["node_modules/brisa/out/cli/build.js", "PROD"],
+        [path.join(outPath, 'cli', 'serve', 'index.js'), `${PORT}`, "PROD"],
         prodOptions,
       );
     }
+
+    // Command: brisa --help
+    else {
+      console.log("Command not found");
+      console.log("Usage: brisa <command> [options]");
+      console.log("Commands:");
+      console.log(" dev           Start development server");
+      console.log(" build         Build for production");
+      console.log(" start         Start production server");
+      console.log("Options:");
+      console.log(" --help        Show help");
+      console.log(
+        " --port        Specify port (applicable for dev and start commands)",
+      );
+      return process.exit(0);
+    }
+  } catch (error) {
+    console.error("Error:", error.message);
+    return process.exit(1);
   }
 
-  // Command: brisa start
-  else if (process.argv[2] === "start") {
-    let PORT = 3000; // default port
+  async function initTauri(options = devOptions, port = 3000) {
+    const tauriConfigPath = path.join(process.cwd(), "src-tauri", "tauri.conf.json");
 
-    for (let i = 3; i < process.argv.length; i++) {
-      switch (process.argv[i]) {
-        case "-p":
-        case "--port":
-          PORT = process.argv[i + 1];
-          i++;
-          break;
-        case "--help":
-          console.log("Usage: brisa start [options]");
-          console.log("Options:");
-          console.log(" -p, --port    Specify port");
-          console.log(" --help        Show help");
-          process.exit(0);
-      }
+    if (!packageJSON?.dependencies?.["@tauri-apps/cli"]) {
+      console.log("Installing @tauri-apps/cli...");
+      cp.spawnSync(BUN_EXEC, ["i", "@tauri-apps/cli"], options);
     }
 
-    spawnSync(
-      BUN_EXEC,
-      ["node_modules/brisa/out/cli/serve/index.js", `${PORT}`, "PROD"],
-      prodOptions,
-    );
-  }
+    if(fs.existsSync(tauriConfigPath)) return
 
-  // Command: brisa --help
-  else {
-    console.log("Command not found");
-    console.log("Usage: brisa <command> [options]");
-    console.log("Commands:");
-    console.log(" dev           Start development server");
-    console.log(" build         Build for production");
-    console.log(" start         Start production server");
-    console.log("Options:");
-    console.log(" --help        Show help");
-    console.log(
-      " --port        Specify port (applicable for dev and start commands)",
-    );
-    process.exit(0);
-  }
-} catch (error) {
-  console.error("Error:", error.message);
-  process.exit(1);
-}
-
-async function initTauri() {
-  if (!packageJSON.dependencies["@tauri-apps/cli"]) {
-    console.log("Installing @tauri-apps/cli...");
-    spawnSync(BUN_EXEC, ["i", "@tauri-apps/cli"], devOptions);
-  }
-
-  if (
-    !fs.existsSync(path.join(process.cwd(), "src-tauri", "tauri.conf.json"))
-  ) {
-    const name = packageJSON.name ?? "my-app";
+    const name = packageJSON?.name ?? "my-app";
     const initTauriCommand = [
       "tauri",
       "init",
@@ -194,7 +193,7 @@ async function initTauri() {
       "-D",
       "../out",
       "--dev-path",
-      "http://localhost:3000",
+      `http://localhost:${port}`,
       "--before-dev-command",
       "bun dev -- -s",
       "--before-build-command",
@@ -202,17 +201,19 @@ async function initTauri() {
     ];
 
     console.log("Initializing Tauri...");
-    spawnSync(BUNX_EXEC, initTauriCommand, devOptions);
+    cp.spawnSync(BUNX_EXEC, initTauriCommand, options);
 
-    const tauriConf = await import(
-      path.join(process.cwd(), "src-tauri", "tauri.conf.json")
-    ).then((m) => m.default);
+    if (!fs.existsSync(tauriConfigPath)) return
+
+    const tauriConf = await import(tauriConfigPath).then((m) => m.default);
 
     // change the bundle identifier in `tauri.conf.json > tauri > bundle > identifier` to `com.${name}`
     tauriConf.tauri.bundle.identifier = `com.${name}`;
     fs.writeFileSync(
-      path.join(process.cwd(), "src-tauri", "tauri.conf.json"),
+      tauriConfigPath,
       JSON.stringify(tauriConf, null, 2),
     );
   }
 }
+
+if (import.meta.main) main();
